feat(List): delete AddTask and replace from List
